safari-600.7-branch/LayoutTests/platform/win/js/dom/global-constructors-attributes-dedicated-worker-expected.txt

Test to ensure that global constructors have the right attributes in the dedicated worker context

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: script-tests/global-constructors-attributes.js
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').value is ArrayBuffer
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ArrayBuffer').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').value is Blob
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').value is DataView
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DataView').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').value is DedicatedWorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').value is EventSource
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').value is Float32Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float32Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').value is Float64Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Float64Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').value is Int16Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int16Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').value is Int32Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int32Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').value is Int8Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Int8Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').value is MessageChannel
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').value is MessageEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').value is Uint16Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint16Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').value is Uint32Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint32Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').value is Uint8Array
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8Array').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').value is Uint8ClampedArray
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Uint8ClampedArray').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').value is WebSocket
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').value is WorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').value is WorkerLocation
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').value is XMLHttpRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').configurable is true
PASS successfullyParsed is true

TEST COMPLETE
